src/requirements.txt
modelcontextprotocol
sqlite-vec
sentence-transformers
beautifulsoup4
markdown
python-frontmatter
fastmcp
gdown

# Development tools
flake8
black
isort
mypy